requirements.txt
Django
Pillow
django-cors-headers
djangorestframework
djangorestframework-simplejwt==4.4.0